pyproject.toml
[tool.poetry]
name = "marshab"
version = "0.1.0"
description = "Mars Habitat Site Selection and Rover Navigation System"
authors = ["MarsHab Development Team"]
readme = "README.md"
packages = [{include = "marshab"}]

[tool.poetry.dependencies]
python = "^3.11"
rasterio = "^1.3"
geopandas = "^0.14"
shapely = "^2.0"
fiona = "^1.9"
pyproj = "^3.6"
spiceypy = "^6.0"
numpy = "^1.26"
scipy = "^1.11"
scikit-image = "^0.22"
pandas = "^2.1"
xarray = "^2023.12"
plotly = "^5.18"
matplotlib = "^3.8"
typer = {extras = ["all"], version = "^0.20"}
rich = "^13.7"
pydantic = "^2.5"
pydantic-settings = "^2.1"
pyyaml = "^6.0"
python-dotenv = "^1.0"
structlog = "^24.1"
fastapi = "^0.104"
uvicorn = {extras = ["standard"], version = "^0.24"}
websockets = "^12.0"
python-multipart = "^0.0.6"
pillow = "^10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-mock = "^3.12"
pytest-asyncio = "^0.21"
httpx = "^0.25"  # Required for FastAPI TestClient
ruff = "^0.1"
mypy = "^1.7"
black = "^23.12"
isort = "^5.13"
ipython = "^8.18"
ipykernel = "^6.29"

[tool.poetry.scripts]
marshab = "marshab.cli:app"
mars = "marshab.cli:mars_app"
marshab-web = "marshab.web.server:main"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=marshab --cov-report=html --cov-report=term"

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

